requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
gunicorn==25.3.0
lxml==6.0.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
python-docx==1.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.12.0
cloudinary
django-cloudinary-storage